backend/requirements.txt
alembic==1.16.5
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==3.1.2
flask-cors==6.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
gunicorn==20.1.0
greenlet==3.2.4
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==2.1.5
numpy==2.3.3
packaging==25.0
pandas==2.3.2
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
reportlab==4.0.7
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.43
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.3